Model Maker: refactor dataloader.gen_datast preprocessing logic
This new refactor assumes that the `preprocess` argument passed in takes an additional `is_training` parameter, which is handy for image preprocessor and potentially other pre-processor as we can squash both pre-processor and data augmentation into one API.

This change updates the internal API usage and updated all the references (image preprocessor only for now). No changes required on the external API.

A potential future improvement that could improve reduce the training throughput is to do preprocessing in batches. The image preprocessing is already capable of doing so. This will require changing the argument `preprocess` to `batch_preprocess` and doing preprocess after batch. But things could get complicated if we want to enable caching in the future.

PiperOrigin-RevId: 342608001