Model Maker audio refactor 2: Lazily load dataset upon gen_dataset is called.
loader._dataset will store a Dataset of (filenames, labels)

when gen_dataset is called, the above raw dataset is translated to the actual spectrum dataset. The number of elements in the two dataset doesn't match. This is possible since Audio dataset is not repeated and task doesn't need the exact training samples count for steps_per_epochs.

Split still works, except that it takes a fraction of the raw dataset, not the spectrum dataset, which we don't know the length of.

PiperOrigin-RevId: 354310582